feat: add cud user operations

+++ src/modules/user/services/user.service.ts
@@ -1,7 +1,13 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { UserRepository } from '../repositories/user.repository';
 import { CreateUserDto } from '../dto/create-user.dto';
 import { isUUID } from '@nestjs/common/utils/is-uuid';
+import { UpdatePasswordDto } from '../dto/update-password.dto';
 
 @Injectable()
 export class UserService {
@@ -12,7 +18,7 @@ export class UserService {
   }
 
   findUserById(id: string) {
-    if (!isUUID(id)) throw new NotFoundException(`Invalid user id ${id}`);
+    if (!isUUID(id)) throw new BadRequestException(`Invalid user id ${id}`);
 
     const user = this.userRepository.findUserById(id);
     if (!user) throw new NotFoundException(`User with id ${id} not found`);
@@ -21,14 +27,37 @@ export class UserService {
   }
 
   createUser(userData: CreateUserDto) {
+    if (!userData.login || !userData.password)
+      throw new BadRequestException('Missing required fields');
+
     return this.userRepository.createUser(userData);
   }
 
-  updateUserPassword(id: string, newPassword: string) {
+  updateUserPassword(
+    id: string,
+    { oldPassword, newPassword }: UpdatePasswordDto,
+  ) {
+    if (!isUUID(id)) throw new BadRequestException(`Invalid user id ${id}`);
+
+    const user = this.userRepository.findUserById(id);
+    if (!user) throw new NotFoundException('User not found');
+
+    if (user.password !== oldPassword)
+      throw new ForbiddenException('Old password is wrong');
+
     return this.userRepository.updateUserPassword(id, newPassword);
   }
 
   deleteUser(id: string) {
+    if (!isUUID(id)) {
+      throw new BadRequestException('Invalid user ID');
+    }
+
+    const result = this.userRepository.deleteUser(id);
+    if (!result) {
+      throw new NotFoundException('User not found');
+    }
+
     return this.userRepository.deleteUser(id);
   }
 }
